+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [6.1.2 (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v6.1.1...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [6.1.1 (2026-01-29)](https://github.com/kdeldycke/meta-package-manager/compare/v6.1.0...v6.1.1)
 
 - Add `--retry-count=3` option to all `choco` invocations.